Allow the use of custom users
- Bugzilla Bug #1540091 - Unable to have non "pkiuser" owned CA
  instance (alee, mharmsen)

Fix nuxwdog to work on all platforms

- dogtagpki Pagure Issue #2874 - nuxwdog won't start on Fedora

(cherry picked from commit 19f04606cb58f7dc6e2872e39f62d3dad427ae22)

Modified systemd invocations in pkispawn to handle nuxwdog

The systemd invocations in pkispawn/pkidestroy did not account for
nuxwdog enabled instances.  This patch allows pkispawn/pkidestroy to
use the right service name if the nuxwdog service unit files exist.

Also modified instance_layout deployment script to delete the right
systemd link.

(cherry picked from commit c7c907c07599ef1d9b52638c25153f7bd82de999)

Allow prompting for token passwords if not present

(cherry picked from commit 6716b82ecc38b23de81c8f0fe18863e1df4bfddb)

Allow instances to be created with custom users

Some folks want to run instances under a different user and
group (ie. not pkiuser).  They may even want a different user for
each instance.  The way to do this in systemd is to create systemd
override files for the specific instance.

The deployment scriptlets have been updated to create (and delete)
these override files.

(cherry picked from commit afe0a2525cace41a1bef2ff7fe0f8f53aa5990e5)

Make sure tomcat is running as pki user with nuxwdog

The nuxwdog process needs to run as a privileged user to be able
to retrieve the passwords from the systemd tty agent in systemctl.
Therefore, the nuxwdog unit file should NOT specify the PKI user
there.

However, we have added an option to nuxwdog to specify the user
in the nuxwdog config file, so that the process that nuxwdog spawns
(ie. tomcat) will run as the specified user.

The code changes in this patch ensure that when the nuxwdog conf
file is created, the user is set correctly as the value of the
variable TOMCAT_USER.

(cherry picked from commit 96e99209b278b5ba380a61486a5b6ce6a87326a0)

More fixes for non-standard users

Needed to fix some python code that was added that works only on Python 3.
The top level directories for the registry should be owned by
root and be world readable/executable so that different users
can read the registry.

cherry-pick from 79e8a8e99a3c9c6d16b63bb5d4b2b18a4b828b11

(cherry picked from commit e37ddfbc8b54e17dbe081221f6c849732a30ebe2)

+++ base/server/python/pki/server/deployment/scriptlets/instance_layout.py
@@ -58,7 +58,8 @@ class PkiScriptlet(pkiscriptlet.AbstractBasePkiScriptlet):
             # Link /etc/pki/<instance>/logging.properties
             # to /usr/share/pki/server/conf/logging.properties.
             deployer.symlink.create(
-                os.path.join(deployer.mdict['pki_source_server_path'], "logging.properties"),
+                os.path.join(deployer.mdict['pki_source_server_path'],
+                             "logging.properties"),
                 os.path.join(deployer.mdict['pki_instance_configuration_path'],
                              "logging.properties"))
 
@@ -143,10 +144,21 @@ class PkiScriptlet(pkiscriptlet.AbstractBasePkiScriptlet):
             deployer.symlink.create(
                 deployer.mdict['pki_tomcat_bin_path'],
                 deployer.mdict['pki_tomcat_bin_link'])
+
+            # create systemd links
             deployer.symlink.create(
                 deployer.mdict['pki_tomcat_systemd'],
                 deployer.mdict['pki_instance_systemd_link'],
                 uid=0, gid=0)
+            user = deployer.mdict['pki_user']
+            group = deployer.mdict['pki_group']
+            if user != 'pkiuser' or group != 'pkiuser':
+                deployer.systemd.set_override(
+                    'Service', 'User', user, 'user.conf')
+                deployer.systemd.set_override(
+                    'Service', 'Group', group, 'user.conf')
+            deployer.systemd.write_overrides()
+            deployer.systemd.daemon_reload()
 
             # establish shared NSS security databases for this instance
             deployer.directory.create(deployer.mdict['pki_database_path'])
@@ -173,7 +185,14 @@ class PkiScriptlet(pkiscriptlet.AbstractBasePkiScriptlet):
         if len(deployer.instance.tomcat_instance_subsystems()) == 0:
 
             # remove Tomcat instance systemd service link
-            deployer.symlink.delete(deployer.mdict['pki_systemd_service_link'])
+            deployer.symlink.delete(deployer.systemd.systemd_link)
+
+            # delete systemd override directories
+            if deployer.directory.exists(deployer.systemd.base_override_dir):
+                deployer.directory.delete(deployer.systemd.base_override_dir)
+            if deployer.directory.exists(deployer.systemd.nuxwdog_override_dir):
+                deployer.directory.delete(deployer.systemd.nuxwdog_override_dir)
+            deployer.systemd.daemon_reload()
 
             # remove Tomcat instance base
             deployer.directory.delete(deployer.mdict['pki_instance_path'])
@@ -192,8 +211,9 @@ class PkiScriptlet(pkiscriptlet.AbstractBasePkiScriptlet):
                 deployer.mdict['pki_instance_registry_path'])
             # remove Tomcat PKI registry (if empty)
             if deployer.instance.tomcat_instances() == 0:
-                deployer.directory.delete(
-                    deployer.mdict['pki_instance_type_registry_path'])
+                if deployer.mdict['pki_root_prefix'] != "":
+                    deployer.directory.delete(
+                        deployer.mdict['pki_instance_type_registry_path'])
 
 
 # Callback only when the /usr/share/pki/server/conf directory
